Add files via upload
This Flask-based web application leverages the power of face recognition to identify individuals in a live video stream. It loads a directory of known faces, compares them to the faces in the real-time video feed, and annotates the video with the names of recognized individuals.  You can populate the images directory with images of known individuals, make sure to save the images with the names of the individuals as this is what is extracted for the labelliing. The application extracts and stores face encodings for each image, recognized faces are labeled with their names, and a bounding box is drawn around them in real-time. If a face matches one in the known faces database, it displays the name; otherwise, it labels the face as "Unknown."